extract validation function to it own folder
