+++ commands/moveRoom.js
@@ -1,10 +1,11 @@
-const {locationState, locationLookUp, roomNameLookup} = require('../helpers/lookUps')
-const {NoRoomSelected, RoomDoesntExistError, MoveRoomError, NotUnlockedError} = require('../errors/roomErrors')
+const {roomNameLookup} = require('../helpers/lookUps')
+const {NotUnlockedError} = require('../errors/roomErrors')
 const {movePlayer} = require('../helpers/movePlayer')
 const {getObjectName} = require('../helpers/getFunctions')
 const {puzzleLookup} = require('../helpers/puzzlesLookup')
 const {printOutput} = require('../helpers/printOutput')
 const {displayRoomPuzzle} = require('../helpers/displayPuzzle')
+const {validateMove} = require('../validation/validateMove')
 
 async function moveRoom(player, targetedRoom) {
   targetedRoom = getObjectName(targetedRoom, roomNameLookup)
@@ -24,37 +25,6 @@ async function moveRoom(player, targetedRoom) {
       throw error
     }
   }
-}
-
-function validateMove(currentRoom, targetedRoom) {
-  if (!targetedRoom && typeof targetedRoom === 'boolean') {
-    throw new NoRoomSelected('No room selected! 🚫')
-  }
-  if (!checkRoomExists(targetedRoom)) {
-    throw new RoomDoesntExistError('Room does not exist! 🚫')
-  }
-  if (!checkValidMove(currentRoom, targetedRoom)) {
-    throw new MoveRoomError("You can't move to this room! 🚫")
-  }
-  if (!solvePuzzle(currentRoom, targetedRoom)) {
-    throw new NotUnlockedError('Please solve the puzzle first! 🧩')
-  }
-}
-
-function checkRoomExists(targetedRoom) {
-  return locationLookUp.hasOwnProperty(targetedRoom)
-}
-
-function checkValidMove(currentRoom, targetedRoom) {
-  return locationState[currentRoom].includes(targetedRoom)
-}
-
-function checkUnlocked(targetedRoom) {
-  return locationLookUp[targetedRoom].isUnlocked
-}
-
-function solvePuzzle(currentRoom, targetedRoom) {
-  return !(checkValidMove(currentRoom, targetedRoom) && !checkUnlocked(targetedRoom))
 }
 
 export {moveRoom}
