+++ commands/showPlayerInventory.js
@@ -1,6 +1,3 @@
-const {EmptyInventory} = require('../errors/itemErrors')
-const {print} = require('../helpers/print')
-
 function showPlayerInventory(player) {
   try {
     validateInventory(player)
@@ -10,13 +7,4 @@ function showPlayerInventory(player) {
   }
 }
 
-function validateInventory(player) {
-  if (!hasItem(player.inventory)) {
-    throw new EmptyInventory('You inventory is empty')
-  }
-}
-
-function hasItem(inventory) {
-  return inventory.length > 0
-}
 exports.showPlayerInventory = showPlayerInventory
