--- cypress/e2e/integration/game_spec.cy.js
@@ -1,1 +1,0 @@
-// game logic tests
