+++ helpers/checkItemExist.js
@@ -1,9 +1,13 @@
 const {itemLookUp} = require('./itemsLookUp')
 
 function checkItemExist(item) {
-  console.log('item', item)
-  console.log('itemLookUp', itemLookUp)
-  console.log('itemLookUp.hasOwnProperty(item)', itemLookUp.hasOwnProperty(item))
-  return itemLookUp.hasOwnProperty(item)
+  if (!item && typeof item === 'boolean') {
+    throw new Error('No item selected! 🚫')
+  }
+  if (!itemLookUp.hasOwnProperty(item)) {
+    throw new Error('Item not found. Please try again. 🔄')
+  }
+  return true
 }
+
 exports.checkItemExist = checkItemExist
